Add contacts.html and links

+++ src/main/java/ca/qc/colval/finalprojectweb/BLL/Models/DTO/ContactDTO.java
@@ -11,5 +11,5 @@ public class ContactDTO {
     private Long contactId;
     private String firstName;
     private String lastName;
-    private int compteId;
+    private Long compteId;
 }
